fix(oidc): improve oidc code structure
